@@ -34,40 +39,51 @@
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "lottie-ios": "^3.4.0",
-    "lottie-react-native": "5.1.6",
+    "lottie-react-native": "7.1.0",
     "moment": "^2.29.4",
     "mongodb-stitch": "^3.18.0",
     "nativewind": "^2.0.11",
-    "react": "18.2.0",
+    "node-cron": "^3.0.3",
+    "node-notifier": "^10.0.1",
+    "node-pushnotifications": "^2.1.0",
+    "react": "18.3.1",
     "react-clock": "^4.5.1",
+    "react-countdown-circle-timer": "^3.2.1",
     "react-i18next": "^13.5.0",
-    "react-native": "0.72.5",
+    "react-native": "0.76.2",
     "react-native-animatable": "^1.3.3",
     "react-native-calendars": "^1.1302.0",
+    "react-native-countdown-circle-timer": "^3.2.1",
     "react-native-dotenv": "^3.4.9",
     "react-native-elements": "^3.4.3",
     "react-native-feather": "^1.1.2",
+    "react-native-gesture-handler": "~2.20.2",
     "react-native-heroicons": "^3.2.0",
     "react-native-image-picker": "^7.0.2",
     "react-native-lightbox": "^0.8.1",
-    "react-native-maps": "^1.7.1",
+    "react-native-linear-gradient": "^2.8.3",
+    "react-native-maps": "1.18.0",
     "react-native-maps-directions": "^1.9.0",
     "react-native-modal": "^13.0.1",
-    "react-native-paper": "^5.11.1",
+    "react-native-paper": "^5.12.3",
     "react-native-progress": "^5.0.0",
+    "react-native-push-notification": "^8.1.1",
     "react-native-qrcode-svg": "^6.2.0",
     "react-native-ratings": "^8.1.0",
-    "react-native-safe-area-context": "4.6.3",
-    "react-native-screens": "~3.22.0",
-    "react-native-snap-carousel": "^3.9.1",
-    "react-native-star-rating": "^1.1.0",
-    "react-native-svg": "13.9.0",
+    "react-native-reanimated": "~3.16.1",
+    "react-native-reanimated-list": "^1.0.0-alpha4",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.1.0",
+    "react-native-snap-carousel": "^1.3.1",
+    "react-native-star-rating": "^1.0.8",
+    "react-native-svg": "15.8.0",
+    "react-native-vector-icons": "^10.0.3",
     "react-query": "^3.39.3",
-    "socket.io-client": "^4.7.2",
+    "socket.io-client": "^4.7.5",
     "tailwindcss-react-native": "^1.7.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
+    "@babel/core": "^7.25.2",
     "gh-pages": "^6.1.1",
     "postcss": "^8.4.31",
     "tailwindcss": "3.3.2"
